requirements.txt
# requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
streamlit==1.30.0
pandas==2.1.3
nltk==3.8.1
transformers==4.38.1
fastapi==0.108.0
uvicorn==0.25.0
scikit-learn==1.3.2
pytest==7.4.3
newspaper3k==0.2.8
matplotlib==3.8.2
gradio==4.12.0
pydantic==2.4.2
google-search-results==2.4.2
huggingface-hub==0.20.2
gTTS==2.4.0
soundfile==0.12.1
lxml>=4.9.3
lxml_html_clean>=0.1.1
newsapi-python==0.2.7
sentencepiece==0.2.0
protobuf==3.20.3 